+++ Runtime/State/EventBuilder.cs
@@ -1,3 +1,4 @@
+using System;
 using System.Collections.Generic;
 using TestingFloor.Internal;
 
@@ -44,11 +45,37 @@ namespace TestingFloor {
             return this;
         }
 
+        public EventBuilder Set(string key, Guid value) {
+            _properties ??= TelemetryQueue.RentProperties();
+            _properties[key] = value;
+            return this;
+        }
+
         public EventBuilder SetIfPresent(string key, string value) {
             if (string.IsNullOrWhiteSpace(value)) return this;
             _properties ??= TelemetryQueue.RentProperties();
             _properties[key] = value;
             return this;
+        }
+
+        public EventBuilder SetIfPresent(string key, int? value) {
+            return value.HasValue ? Set(key, value.Value) : this;
+        }
+
+        public EventBuilder SetIfPresent(string key, long? value) {
+            return value.HasValue ? Set(key, value.Value) : this;
+        }
+
+        public EventBuilder SetIfPresent(string key, float? value) {
+            return value.HasValue ? Set(key, value.Value) : this;
+        }
+
+        public EventBuilder SetIfPresent(string key, double? value) {
+            return value.HasValue ? Set(key, value.Value) : this;
+        }
+
+        public EventBuilder SetIfPresent(string key, bool? value) {
+            return value.HasValue ? Set(key, value.Value) : this;
         }
 
         public void Send() {